feat(pdftk): Extract a page range of a PDF
To extract a page range of a PDF you could print to file using `evince`
selecting a page range. The drawback of this method is that the text might be
converted to images an thus no longer be searchable. A faster alternative that
also keeps the text intact is using `pdftk`.