feat(fastapi): add route to other module functions
There are two main ways of using this module, with the decorator
`@app.get("{{route}}")` or with
`FastAPI().add_route("{{route}}", {{function}})`.

The first alternative is by far the most common, but if you want to separate
the functions from the HTTP API entrypoint, you have to use the second option.

The advantage of this alternative is that we can import the function
from another module.