feat(tdd): exploratory coding (spiking)
When learning a new tool or exploring a new possible solution, it’s often
appropriate to leave the rigorous TDD process to one side, and build a little
prototype without tests, or perhaps with very few tests. The goat doesn't mind
looking the other way for a bit.

For de-spiking, rewrite your prototype code using TDD.