feat(uptimed): uptimed
[`uptimed`](https://github.com/rpodgorny/uptimed) is an uptime record daemon
keeping track of the highest uptimes a computer system ever had. It uses the
system boot time to keep sessions apart from each other. `uptimed` comes with
a console front-end to parse the records, which can also easily be used to
show your records on a web page.

To set up the program just install it with your package manager and it will
start running in the background by default and keeping track of your uptime.

To check the stats run `uprecords`.