feat(the_problem_of_political_authority): The Problem of Political Authority
Modern states commonly deploy coercion in a wide array of circumstances in
which the resort to force would clearly be wrong for any private agent. What
entitles the state to behave in this manner? And why should citizens obey its
commands? This book examines theories of political authority, from the social
contract theory, to theories of democratic authorization, to fairness- and
consequence-based theories. Ultimately, no theory of authority succeeds, and
thus no government has the kind of authority often ascribed to governments.

The author goes on to discuss how voluntary and competitive institutions could
provide the central goods for the sake of which the state is often deemed
necessary, including law, protection from private criminals, and national
security. An orderly and livable society thus does not require acquiescence in
the illusion of political authority.