feat(t14_amd_gen1): sticky touchpad after suspend
After suspension, the touchpad behaves weird: can't scroll, and acts as if you
were clicking and dragging (sticky). The TrackPoint works correctly though.

A first solution was to disable it and just use the TrackPoint. This is done
as follows:

```bash
xinput list # look for SynPS/2 Synaptics TouchPad
xinput set-prop <number> "Device Enabled" 0
```

This allows you to use the computer without the sticky mouse annoyance but
doesn't actually solve the problem. For that, the solution that worked for me
is removing the `psmouse` kernel module before suspending and then reloading it
after waking up from suspension. You can first try to do this manualy and then
if it works create this script in `/lib/systemd/system-sleep/psmouse`:

```bash

case $1/$2 in
  pre/*)
    echo "Going to $2..."
    modprobe -r psmouse
    ;;
  post/*)
    echo "Waking up from $2..."
    sleep 2
    modprobe psmouse
    ;;
esac
```